project/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-js-asset==1.2.2
django-recaptcha3==0.2.0
django-robots==4.0
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.9
